% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_alphanumeric_lowercase.R
\name{df_to_alphanumeric_lowercase}
\alias{df_to_alphanumeric_lowercase}
\title{set string to lower case and replace everything that is not alphanumeric or underscore by a dot "."}
\usage{
df_to_alphanumeric_lowercase(x, convert_content = T,
  convert.colnames = T)
}
\arguments{
\item{x}{a dataframe containing strings or vectors to be converted}

\item{convert.colnames}{T/F on whether column names should be converted to lowercase. Defaults to True}
}
\value{
vector of strings with characters replaced
}
\description{
set string to lower case and replace everything that is not alphanumeric or underscore by a dot "."
}
