% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_alphanumeric_lowercase.R
\name{to_alphanumeric_lowercase}
\alias{to_alphanumeric_lowercase}
\title{set string to lower case and replace everything that is not alphanumeric or underscore by a dot "."}
\usage{
to_alphanumeric_lowercase(x)
}
\arguments{
\item{x}{string or vector of strings}
}
\value{
vector of strings with characters replaced
}
\description{
set string to lower case and replace everything that is not alphanumeric or underscore by a dot "."
}
